products/pyproject.toml
[tool.poetry]
name = "products"
version = "0.1.0"
description = ""
authors = ["esilean <le.bevilaqua@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.15"
psycopg2-binary = "^2.9.6"
pytest = "^7.3.1"
alembic = "^1.11.1"
python-decouple = "^3.8"
httpx = "^0.24.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
